staticanalysis/bot: Move workflows in module
