+++ src/staticanalysis/bot/static_analysis_bot/workflows/local.py
@@ -7,15 +7,11 @@ from __future__ import absolute_import
 
 import os
 import shutil
-from datetime import datetime
-from datetime import timedelta
 
 import hglib
 
 from cli_common.command import run_check
 from cli_common.log import get_logger
-from cli_common.phabricator import PhabricatorAPI
-from cli_common.taskcluster import TASKCLUSTER_DATE_FORMAT
 from static_analysis_bot import CLANG_FORMAT
 from static_analysis_bot import CLANG_TIDY
 from static_analysis_bot import COVERAGE
@@ -36,42 +32,15 @@ from static_analysis_bot.coverity.coverity import Coverity
 from static_analysis_bot.infer import setup as setup_infer
 from static_analysis_bot.infer.infer import Infer
 from static_analysis_bot.lint import MozLint
-from static_analysis_bot.report.debug import DebugReporter
-from static_analysis_bot.revisions import Revision
+from static_analysis_bot.workflows.base import Workflow
 
 logger = get_logger(__name__)
 
-TASKCLUSTER_NAMESPACE = 'project.releng.services.project.{channel}.static_analysis_bot.{name}'
-TASKCLUSTER_INDEX_TTL = 7  # in days
 
-
-class Workflow(object):
+class LocalWorkflow(Workflow):
     '''
-    Static analysis workflow
+    Run analyers in current task
     '''
-    def __init__(self, reporters, analyzers, index_service, queue_service, phabricator_api):
-        assert isinstance(analyzers, list)
-        assert len(analyzers) > 0, \
-            'No analyzers specified, will not run.'
-        self.analyzers = analyzers
-        assert 'MOZCONFIG' in os.environ, \
-            'Missing MOZCONFIG in environment'
-
-        # Use share phabricator API client
-        assert isinstance(phabricator_api, PhabricatorAPI)
-        self.phabricator = phabricator_api
-
-        # Load reporters to use
-        self.reporters = reporters
-        if not self.reporters:
-            logger.warn('No reporters configured, this analysis will not be published')
-
-        # Always add debug reporter and Diff reporter
-        self.reporters['debug'] = DebugReporter(output_dir=settings.taskcluster.results_dir)
-
-        # Use TC services client
-        self.index_service = index_service
-        self.queue_service = queue_service
 
     @stats.api.timed('runtime.clone')
     def clone(self):
@@ -328,34 +297,3 @@ class Workflow(object):
                 issues += analyzer_issues
 
         return issues
-
-    def index(self, revision, **kwargs):
-        '''
-        Index current task on Taskcluster index
-        '''
-        assert isinstance(revision, Revision)
-
-        if settings.taskcluster.local:
-            logger.info('Skipping taskcluster indexing', rev=str(revision), **kwargs)
-            return
-
-        # Build payload
-        payload = revision.as_dict()
-        payload.update(kwargs)
-
-        # Always add the indexing
-        now = datetime.utcnow()
-        payload['indexed'] = now.strftime(TASKCLUSTER_DATE_FORMAT)
-
-        # Index for all required namespaces
-        for name in revision.namespaces:
-            namespace = TASKCLUSTER_NAMESPACE.format(channel=settings.app_channel, name=name)
-            self.index_service.insertTask(
-                namespace,
-                {
-                    'taskId': settings.taskcluster.task_id,
-                    'rank': 0,
-                    'data': payload,
-                    'expires': (now + timedelta(days=TASKCLUSTER_INDEX_TTL)).strftime(TASKCLUSTER_DATE_FORMAT),
-                }
-            )
